Fixed a bug
There is a bug where if someone removes a block the block under it
gets the value of the above just removed block, This is because react doesn't
re-render the block it just gives the already there block that is "removed" the
data of the one under it.
This is a problem because the input fields will only update on the first
props recieved. This has as result that when a block updates the input
will just stay there and recieve a new intial value but the input
doesn't do anything with it.